backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
tensorflow==2.15.0
scikit-learn==1.4.0
joblib==1.3.2
numpy==1.26.3
Pillow==10.2.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
